Add evaluate_model function to base model class and update some documentation.
