+++ garisom_tools/model/sperry.py
@@ -1,17 +1,11 @@
 # Basic data utils
 import pandas as pd
-import numpy as np
 from typing import Any
 
 # For model evaluation
 import os
 import subprocess
 from tempfile import TemporaryDirectory
-from datetime import datetime
-
-# Optimizer stuff
-from garisom_tools.config import MetricConfig
-from garisom_tools.utils.results import EvalResults
 
 # Parallel runs
 from tqdm import tqdm
@@ -410,118 +404,3 @@ class SperryModel(Model):
         out_file = pd.read_csv(output_file)
 
         return out_file
-
-    @classmethod
-    def evaluate_model(
-        cls,
-        output,
-        ground,
-        metric_config: MetricConfig,
-        start_date: datetime,
-        end_date: datetime
-    ) -> EvalResults:
-        """
-        Evaluate model predictions against ground truth observations.
-
-        This method computes multiple evaluation metrics by comparing model output
-        to observed data over a specified time period. It handles data alignment,
-        missing values, and multiple output variables.
-
-        Args:
-            output (pd.DataFrame | None): Model output DataFrame with timestep data.
-                Must contain columns matching metric_config output names.
-                If None, returns penalty values for all metrics.
-            ground (pd.DataFrame): Ground truth observations DataFrame.
-                Must contain 'julian-day' column and columns matching output variables.
-            metric_config (MetricConfig): Configuration specifying which metrics to compute.
-                Contains lists of metrics, their target output variables, and optimization modes.
-            start_day (int): Julian day to start evaluation period (inclusive).
-            end_day (int): Julian day to end evaluation period (inclusive).
-
-        Returns:
-            EvalResults: Dictionary mapping metric names to computed values.
-                For minimization metrics, higher values indicate worse performance.
-                For maximization metrics, higher values indicate better performance.
-
-        Example:
-            ```python
-            from garisom_tools.config import MetricConfig
-
-            # Setup metric configuration
-            metric_config = MetricConfig.from_dict({
-                'metrics': ['rmse', 'r2', 'nse'],
-                'modes': ['min', 'max', 'max'],
-                'params': ['leaftemp', 'leaftemp', 'E-MD']
-            })
-
-            # Load ground truth data
-            ground_truth = pd.read_csv("observations.csv")  # Must have 'julian-day' column
-
-            # Evaluate model output
-            model_output = SperryModel.run(...)
-            evaluation = SperryModel.evaluate_model(
-                output=model_output,
-                ground=ground_truth,
-                metric_config=metric_config,
-                start_date=datetime(2023, 07, 23)
-                end_date=243
-            )
-
-            print(f"RMSE: {evaluation['rmse']:.3f}")
-            print(f"R²: {evaluation['r2']:.3f}")
-            print(f"NSE: {evaluation['nse']:.3f}")
-            ```
-
-        Note:
-            - Data is filtered by date before metric calculation
-            - Missing values (NaN) are automatically excluded from ground truth
-            - Model predictions are aligned with ground truth observations by index
-            - Failed model runs (output=None) receive penalty values (1e20 for min, -1e20 for max)
-            - Supports multiple metrics on the same output variable using suffix notation (e.g., 'var.A', 'var.B')
-        """
-        out_names = [metric.output_name for metric in metric_config.metrics]
-        pred = output[out_names].to_numpy(dtype=float) if output is not None else None
-
-        metrics = metric_config.metrics
-        modes = metric_config.modes
-
-        errors: dict[str, np.typing.ArrayLike] = {}
-        for idx, (metric, mode) in enumerate(zip(metrics, modes)):
-
-            output_name = metric.output_name
-            optim_name = metric.name
-            eval_func = metric.func
-
-            if pred is None or eval_func is None:
-                err = 1e20 if mode == 'min' else -1e20 if mode == "max" else 0
-            else:
-                # Filter ground data based on year and julian-day, then drop NaN values
-                start_year, start_day = start_date.year, start_date.timetuple().tm_yday
-                end_year, end_day = end_date.year, end_date.timetuple().tm_yday
-
-                if start_year == end_year:
-                    mask = (
-                        (ground['year'] == start_year) &
-                        (ground['julian-day'] >= start_day) &
-                        (ground['julian-day'] <= end_day)
-                    )
-                else:
-                    mask = (
-                        ((ground['year'] == start_year) & (ground['julian-day'] >= start_day)) |
-                        ((ground['year'] > start_year) & (ground['year'] < end_year)) |
-                        ((ground['year'] == end_year) & (ground['julian-day'] <= end_day))
-                    )
-                col_ground = ground[mask][output_name].dropna()
-
-                ground_values = np.array([col_ground.to_numpy()]).squeeze(axis=0)
-
-                # Align predictions with the filtered ground data
-                col_pred = pred[:, idx]  # (T)
-                col_pred = pd.DataFrame(col_pred)
-                pred_values = col_pred.loc[col_ground.index].T.to_numpy().squeeze(axis=0)
-
-                err = eval_func(ground_values, pred_values)
-
-            errors[optim_name] = err
-
-        return errors
